+++ hammurabi/law.py
@@ -10,7 +10,7 @@ will be aborted and the original exception will be re-raised.
 """
 
 import logging
-from typing import Iterable, List, Union
+from typing import Iterable, List, Tuple, Union
 
 from hammurabi.config import config
 from hammurabi.exceptions import AbortLawError
@@ -63,10 +63,54 @@ class Law(GitMixin):
         self.name = name.strip()
         self.description = full_strip(description)
         self.rules: Iterable[Rule] = tuple()
-        self.failed_rules: Iterable[Rule] = tuple()
+        self._failed_rules: Union[Tuple[()], Tuple[Rule]] = ()
 
         for rule in rules:
             self.rules += (rule,)
+
+    @property
+    def passed_rules(self) -> Tuple[Rule, ...]:
+        """
+        Return the rules which did modifications and not failed.
+
+        :return: Return the passed rules
+        :rtype: Tuple[Rule, ...]
+        """
+
+        return tuple(r for r in self.rules if r.made_changes)
+
+    @property
+    def failed_rules(self) -> Union[Tuple[()], Tuple[Rule]]:
+        """
+        Return the rules which did modifications and failed.
+
+        :return: Return the failed rules
+        :rtype: Union[Tuple[()], Tuple[Rule]]
+        """
+
+        return self._failed_rules
+
+    @property
+    def skipped_rules(self) -> Tuple[Rule, ...]:
+        """
+        Return the rules which neither modified the code nor failed.
+
+        :return: Return the skipped rules
+        :rtype: Tuple[Rule, ...]
+        """
+
+        def is_skipped(rule) -> bool:
+            """
+            Return the evaluation if the rule is skipped or not.
+
+            :return: Evaluation if the rule is skipped
+            :rtype: bool
+            """
+            not_passed = rule not in self.passed_rules
+            not_failed = rule not in self.failed_rules
+            return not_passed and not_failed
+
+        return tuple(r for r in self.rules if is_skipped(r))
 
     @property
     def documentation(self) -> str:
@@ -100,8 +144,8 @@ class Law(GitMixin):
 
     def commit(self):
         """
-        Commit the changes made by registered rules
-        and add a meaningful commit message.
+        Commit the changes made by registered rules and add a
+        meaningful commit message.
 
         Example commit message:
 
@@ -173,7 +217,7 @@ class Law(GitMixin):
                 self.__execute_rule(rule)
             except AbortLawError as exc:
                 logging.error(str(exc))
-                self.failed_rules += (rule,)
+                self._failed_rules += (rule,)
 
                 if config.settings.rule_can_abort:
                     raise exc